fix: Remove router link from save note FAB
The save note FAB on the add note screen had a router link that would
open the add task screen and then pop that, then go back to the previous
screen. This would show the contents of the old note that was being
saved when going to add a new note. This should fix that bug.

Signed-off-by: Amogh Harish <amogh@aharish.xyz>