fix(core): Fix status picker in task list failing
Picking a status from the status picker in the task list would do
nothing. This was because it would fail silently since the TaskItem
component wasn't updated to use the SQLite adapter on Capacitor instead
of IndexedDB. This commit fixes the regression, and status updates from
the task list should work as expected now.

Signed-off-by: Amogh Harish <amogh@aharish.xyz>