feat(ux): Replace WYSIWYG editor as default in notes
When viewing a note, make the new WYSIWYG / Rich editor the default
experience, instead of having separate screens for viewing the rendered
note, and editing the said note. This will make for a more seamless UX.

As a part of this, save the note on update, by default debouncing the
input so that unnecessary DB operations are avoided.

Signed-off-by: Amogh Harish <amogh@aharish.xyz>