feat(core): Add experimental desktop support
Integrate electron-vite to target the Electron platform to provide
experimental desktop support.

Currently, electron-vite is configured to output built source files, and
electron-builder is only configured to generate an AppImage for Linux.

Build artifacts won't be uploaded or processed by the CI/CD pipeline
until more concrete support for the multi-pane UI is added.

The Linux AppImage configuration is only temporary and will be switched
to a Flatpak build once the multi-pane UI is up to par.

Signed-off-by: Amogh Harish <amogh@aharish.xyz>