feat(ux): More nuanced use of due & scheduled date
Make more nuanced use of the due and scheduled dates. Added support on
the Today screen to order tasks and overdue tasks by the due date field
instead of using the scheduled date field. Add list separation headers
for better UX, and also show tasks that have been started in the past,
but have no due date set in a separate list on the Today screen.

Signed-off-by: Amogh Harish <amogh@aharish.xyz>