fix(core): Move index creation after deviceready
On Capacitor, ensure that the index creation happens only after the
deviceready event from Cordova has fired so that we know that the DB is
ready for use. On non-Capacitor platforms, just fire it as soon as the
PouchDB plugin is setup.

This should fix the issue of the Android app not loading data upon cold
launch at times.

Signed-off-by: Amogh Harish <amogh@aharish.xyz>