fix(core): Call ddoc, index creation when required
In the Projects component, call the ddoc for project related stuff only
when required instead of calling it at app launch; this will ensure
better performance without impacting user experience without need.

Also fetch the project progress before fetching the list of projects.

This fixes an issue where adding a task to an empty and new project,
and then returning to the projects screen would crash the app because
the project would not be fetched yet.

Signed-off-by: Amogh Harish <amogh@aharish.xyz>